userii pot sa vada proiectele + stergerea userilor din proiecte

+++ Taskify/Controllers/ProjectsController.cs
@@ -28,11 +28,13 @@ namespace Taskify.Controllers
             _roleManager = roleManager;
         }
 
-        [Authorize(Roles = "User,Editor,Admin")]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Index()
         {
             var userid = _userManager.GetUserId(User);
-            var projects = db.Projects.Include("User").Where(proj => proj.UserId == userid);
+            var projs = db.UserProjects.Where(userpr => userpr.UserId == userid).Select(c => c.ProjectId);
+            List<int?> proj_ids = projs.ToList();
+            var projects = db.Projects.Include("User").Where(proj => proj_ids.Contains(proj.Id));
             ViewBag.Projects = projects;
             if (TempData.ContainsKey("message"))
             {
@@ -54,6 +56,7 @@ namespace Taskify.Controllers
             return View();
         }
 
+        [Authorize(Roles = "User,Admin")]
         public IActionResult New()
         {
             Project project = new Project();
@@ -82,13 +85,15 @@ namespace Taskify.Controllers
                 return View(project);
             }
         }
-        [Authorize(Roles = "User,Editor,Admin")]
+        [Authorize(Roles = "User,Admin")]
 
         public IActionResult Show(int id)
         {
             var userid = _userManager.GetUserId(User);
+            var users = db.UserProjects.Where(userpr => userpr.ProjectId == id).Select(c => c.UserId);
+            List<string?> user_ids = users.ToList();
             var project = db.Projects.Include("Tasks.User").Include("User").Where(proj => proj.Id == id).First();
-            if (userid == project.UserId || User.IsInRole("Admin"))
+            if (user_ids.Contains(userid) || User.IsInRole("Admin"))
             {
                 return View(project);
             }
@@ -119,56 +124,78 @@ namespace Taskify.Controllers
                 return View(project);
             }
         }
-
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Edit(int id)
         {
-
+            var userid = _userManager.GetUserId(User);
             Project project = db.Projects.Include("Tasks")
                                         .Where(proj => proj.Id == id)
                                         .First();
-
-
-            return View(project);
+            if (project.UserId == userid || User.IsInRole("Admin")) 
+            {
+                return View(project);
+            }
+            else{
+                TempData["message"] = "Error! Nu ai acces";
+                ViewBag.Message = TempData["message"];
+                return Redirect("/Show/" + id);
+            }
 
         }
 
         [HttpPost]
         public IActionResult Edit(int id, Project requestProject)
-        {
+        {   
             Project project = db.Projects.Find(id);
-
-            if (ModelState.IsValid)
+            if(project == null)
             {
-                project.Title = requestProject.Title;
-                project.Description = requestProject.Description;
-                TempData["message"] = "Proiectul a fost modificat";
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                TempData["message"] = "Database error!";
+                return View(requestProject);
+
             }
             else
             {
-                return View(requestProject);
+                if (ModelState.IsValid)
+                {
+                    project.Title = requestProject.Title;
+                    project.Description = requestProject.Description;
+                    TempData["message"] = "Proiectul a fost modificat";
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return View(requestProject);
+                }
             }
+            
         }
 
         [HttpPost]
         public ActionResult Delete(int id)
         {
             Project project = db.Projects.Find(id);
-            db.Projects.Remove(project);
-            db.SaveChanges();
-            TempData["message"] = "Project has been deleted.";
+            if(project == null)
+            {
+                TempData["message"] = "Database error!";
+            }
+            else
+            {
+                db.Projects.Remove(project);
+                db.SaveChanges();
+                TempData["message"] = "Project has been deleted.";
+            }
             return RedirectToAction("Index");
         }
 
-        [Authorize(Roles = "User,Editor,Admin")]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Users(int id)
         {
             var userid = _userManager.GetUserId(User);
             ///de verificat ca userul in sesiune sa fie in cadrul proiectului
             var users_search = db.Users.Where(a => 1 == 0);
             var users = db.UserProjects.Include("User").Where(user => user.ProjectId == id);
-            List<string> users_ids = users.Select(c => c.UserId).ToList();
+            List<string?> users_ids = users.Select(c => c.UserId).ToList();
             var search = "";
             // MOTOR DE CAUTARE
             if (Convert.ToString(HttpContext.Request.Query["search"]) != null)
@@ -194,20 +221,56 @@ namespace Taskify.Controllers
         {
             var userid = _userManager.GetUserId(User);
             var project = db.Projects.Find(id);
-            if (userid == project.UserId || User.IsInRole("Admin"))
+            if (project == null)
             {
-                UserProject userProject = new UserProject();
-                userProject.ProjectId = id;
-                userProject.UserId = requestUser.UserId;
-                db.UserProjects.Add(userProject);
-                db.SaveChanges();
-                return Redirect("/Projects/Users/" + project.Id);
+                TempData["message"] = "Database error!";
+                return RedirectToAction("Index");
             }
             else
             {
-                TempData["message"] = "Error! Nu ai acces";
-                ViewBag.Message = TempData["message"];
-                return RedirectToAction("Index");
+                if (userid == project.UserId || User.IsInRole("Admin"))
+                {
+                    UserProject userProject = new UserProject();
+                    userProject.ProjectId = id;
+                    userProject.UserId = requestUser.UserId;
+                    db.UserProjects.Add(userProject);
+                    db.SaveChanges();
+                    return Redirect("/Projects/Users/" + project.Id);
+                }
+                else
+                {
+                    TempData["message"] = "Error! Nu ai acces";
+                    ViewBag.Message = TempData["message"];
+                    return RedirectToAction("Index");
+                }
+            }
+        }
+
+        [HttpPost]
+        public ActionResult UserDelete(int id, string rmvuser, int rmvproject)
+        {
+            UserProject user_project = db.UserProjects.Find(id, rmvuser, rmvproject);
+            var project_id = user_project.ProjectId;
+            if (user_project == null)
+            {
+                TempData["message"] = "Database error!";
+                return Redirect("/Projects/Users/" + project_id);
+            }
+            else
+            {
+                db.UserProjects.Remove(user_project);
+                db.SaveChanges();
+                if(user_project.UserId == _userManager.GetUserId(User))
+                {
+                    TempData["message"] = "You have left the project.";
+                    return RedirectToAction("Index");
+
+                }
+                else
+                {
+                    TempData["message"] = "User has been removed.";
+                    return Redirect("/Projects/Users/" + project_id);
+                }
             }
         }
     }
